+++ packages/swapkit/helpers/src/helpers/request.ts
@@ -1,13 +1,12 @@
 import type { Options } from 'ky';
 import ky from 'ky';
 
-const kyClient = ky.create({
-  headers: {
-    referrer: 'https://sk.thorswap.net',
-    referer: 'https://sk.thorswap.net',
-    'Content-Type': 'application/json',
-  },
-});
+const headers =
+  typeof window !== 'undefined'
+    ? {}
+    : { referrer: 'https://sk.thorswap.net', referer: 'https://sk.thorswap.net' };
+
+const kyClient = ky.create({ headers });
 
 export const RequestClient = {
   get: <T>(url: string | URL | Request, options?: Options) => kyClient.get(url, options).json<T>(),