+++ packages/swapkit/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swapkit/api
 
+## 1.0.0-rc.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/types@1.0.0-rc.7
+
 ## 1.0.0-rc.6
 
 ### Patch Changes