+++ packages/swapkit/sdk/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @swapkit/sdk
 
+## 1.0.0-rc.23
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.18
+  - @swapkit/types@1.0.0-rc.7
+  - @swapkit/core@1.0.0-rc.21
+  - @swapkit/wallet-keplr@1.0.0-rc.18
+  - @swapkit/wallet-keystore@1.0.0-rc.22
+  - @swapkit/wallet-ledger@1.0.0-rc.20
+  - @swapkit/wallet-okx@1.0.0-rc.20
+  - @swapkit/wallet-wc@1.0.0-rc.19
+  - @swapkit/wallet-xdefi@1.0.0-rc.20
+  - @swapkit/toolbox-evm@1.0.0-rc.17
+  - @swapkit/toolbox-utxo@1.0.0-rc.17
+  - @swapkit/wallet-trezor@1.0.0-rc.18
+  - @swapkit/api@1.0.0-rc.7
+  - @swapkit/wallet-evm-extensions@1.0.0-rc.17
+
 ## 1.0.0-rc.22
 
 ### Patch Changes