+++ packages/toolboxes/cosmos/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @swapkit/toolbox-cosmos
 
+## 1.0.0-rc.18
+
+### Patch Changes
+
+- Use referer from window when needed
+
+- Updated dependencies []:
+  - @swapkit/helpers@1.0.0-rc.16
+  - @swapkit/types@1.0.0-rc.7
+  - @swapkit/api@1.0.0-rc.7
+
 ## 1.0.0-rc.17
 
 ### Patch Changes