+++ packages/wallets/keplr/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @swapkit/wallet-keplr
 
+## 1.0.0-rc.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.18
+  - @swapkit/helpers@1.0.0-rc.16
+  - @swapkit/types@1.0.0-rc.7
+
 ## 1.0.0-rc.17
 
 ### Patch Changes