+++ packages/wallets/keystore/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @swapkit/wallet-keystore
 
+## 1.0.0-rc.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.18
+  - @swapkit/helpers@1.0.0-rc.16
+  - @swapkit/types@1.0.0-rc.7
+  - @swapkit/toolbox-evm@1.0.0-rc.17
+  - @swapkit/toolbox-utxo@1.0.0-rc.17
+
 ## 1.0.0-rc.21
 
 ### Patch Changes