+++ packages/wallets/xdefi/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @swapkit/wallet-xdefi
 
+## 1.0.0-rc.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.18
+  - @swapkit/helpers@1.0.0-rc.16
+  - @swapkit/types@1.0.0-rc.7
+  - @swapkit/toolbox-evm@1.0.0-rc.17
+  - @swapkit/toolbox-utxo@1.0.0-rc.17
+
 ## 1.0.0-rc.19
 
 ### Patch Changes